base: Add database instance as an attribute of the user and group classes
Dynamically construct derived user and group classes at the point that
the database instance is instantiated, and attach the database to each
of these classes (via a weak reference proxy, to avoid circular
reference problems).

This allows the database to be omitted as a parameter when
constructing user or group instances, and also allows user or group
class methods to access the database configuration.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>