base: Define "name" attribute as the canonical user or group name
In most databases, the natural lookup key is expected to be the common
user (or group) name as synchronised between databases.  This will not
be true in all cases: for example, MediaWiki requires user names to be
converted to title case.

Create a separate "name" attribute to allow for a distinction between
the lookup key (which should ideally closely resemble a value existent
within the database in question) and the canonical user or group name.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>