mediawiki: Convert user names to title case
MediaWiki user names must be valid page names and so must be in title
case.  This conflicts with typical username conventions.

Provide a "title_case" configuration option for MediaWiki databases
(defaulting to True), which will cause usernames to be converted to
and from title case as needed.

Note that if conversion is enabled, then external usernames must begin
with a lower-case character; this ensures that a case collision cannot
occur.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>