base: Insist that each database entry has a permanent UUID
Synchronization will require the ability to permanently identify a
database entry across renames and deletions.  Insist that each entry
has a "uuid" attribute to act as a permanent identifier.

For SQL databases, default to constructing a per-row UUID based upon
the table name and the primary key values, on the assumption that the
primary key is usually an immutable identifier for the row.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>