ldap: Use RFC 4533 (syncrepl) to watch for database modifications
LDAP persistent search provides no indication of the transition
between the initial content search and the subsequent updates.  Switch
to using syncrepl, which provides this indication.

Note that there are known problems when operating with FreeIPA as a
source database: the intermediate result indicating the transition is
missing its syncInfoMessage, and entries returned from a resumed
search are missing all operational attributes (including nsuniqueid).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>