ldap: Work around bugs in ldap.syncrepl.SyncInfoMessage
The upstream version of ldap.syncrepl.SyncInfoMessage erroneously
detects a refreshDelete as a refreshPresent.  This appears to be due
to misuse of the underlying pyasn1 API, though the precise root cause
is unclear.

Temporarily work around the problem by reimplementing SyncInfoMessage,
using the ASN.1 types as defined in ldap.syncrepl.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>