ldap: Match behaviour of upstream ldap.syncrepl.SyncInfoMessage
The upstream SyncInfoMessage constructs a dictionary for the message
values, but this is not the same as the simple dictionary that can be
constructed directly from the ASN.1 value.

Match the (arguably suboptimal) format as constructed by the upstream
ldap.syncrepl.SyncInfoMessage, to allow for the fix to be upstreamed.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>