ldap: Clarify usage of refreshDelete and refreshDeletes
The "delete" part of "refreshDelete" or "refreshDeletes" generally
indicates that a delete phase has taken place, i.e. that an explicit
list of entries to be deleted has been provided.  It therefore
indicates that unmentioned entries should not be automatically
deleted.

Rename the RefreshComplete "delete" parameter to "autodelete" to avoid
confusion with the LDAP "delete" concept (which is essentially the
inverse of "autodelete unmentioned entries").

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>