ldap: Allow EntryUUID attribute to be populated from syncStateControl
The 389-ds-base syncrepl implementation ignores the requested
attribute list for search results generated as part of a content
refresh.  The entry UUID is therefore missing since, as an operational
attribute, it is not part of the default attribute list.

Work around this problem by populating the "uuid" attribute from the
entryUUID in the syncStateControl if needed.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>