ldap: Construct SyncId list for easier debugging
A DeletedSyncIds or UnchangedSyncIds object may be constructed with
the SyncId list being a generator expression (or any other iterable).
Given that the underlying syncUUIDs string list has already been
constructed in full, little is saved by using a generator expression.

Switch to constructing the SyncId list in full in order to provide
more meaningful output when __repr__() is called on the resulting
DeletedSyncIds or UnchangedSyncIds object.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>