base: Refactor to support explicit lookups by SyncId or matching entry
Use explicit methods for looking up existing entries, rather than
performing an implicit lookup as part of the entry constructor.
Redefine "key" as being "a suitable unique value that can be passed to
the find() method", convert the database user() and group() methods to
become mere shortcuts for .User.find() and .Group.find(), and update
the user and group __repr__() methods to reflect this potential usage.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>