mediawiki: Store UUIDs as strings rather than raw binary
MySQL has no UUID column type.  Storing UUIDs as BINARY(16) works well
but results in raw binary values appearing in unwanted places such as
the query log, the mysql client terminal output, etc.

Since the SyncID column is not the primary key, and since user
definitions are expected to be modified infrequently, there is a
relatively low impact from switching to use a string representation of
the UUID.  Do so in order to minimise the external annoyances.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>