mediawiki: Generalise concept of a dummy group
The MediaWiki user database has no separate concept of the existence
of a group: a group exists only by virtue of being mentioned by name
(possibly repeatedly) in the memberships table.

It is possible that other user databases may suffer from a similar
lack of database normalisation.  Generalise the concept of a dummy
group to allow for its reuse if needed.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>