base: Refactor bulk deletion
Bulk deletion is difficult to optimise for SQL databases when multiple
tables are involved.  In the most common case, the number of records
to be bulk deleted will be zero.  Optimise for this common case by
changing the internal API so that writable user databases must support
bulk retrieval by synchronization identifier and deletion of
individual records, instead of supporting bulk deletion by
synchronization identifier.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>