sql: Allow syncid column to be stored in a separate table
We currently maintain syncid information in an additional column which
is created automatically as part of the database preparation step.
This may potentially interfere with the application's own upgrade or
other schema maintenance processes.

Maintaining syncid information in a separate table is likely to be
less disruptive.  Databases often contain additional tables (such as
system tables); it should therefore be less intrusive for us to create
an additional table than to create an additional column within an
existing table.

Using a separate table also provides a viable path for undoing the
database preparation step, either automatically or by direct
manipulation of the destination database.

Allow the syncid column to be implemented using a SQLAlchemy
association proxy.  The default query pattern generated for
association proxies tends to be somewhat inefficient; we optimise by
searching directly on the proxied table instead if applicable.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>