sync: Gracefully handle multi-valued to single-valued attribute mappings
There will be situations in which a multi-valued attribute with
multiple values present needs to be mapped to a single-valued
attribute.  For example: a user may have two e-mail addresses within
an LDAP directory, but a destination database may provide only a
single e-mail address column.

We currently treat this as a failure and raise an exception.  This is
likely to be unhelpful in practice, since it will lead to situations
in which synchronization of the user database suddenly stops working
when a user innocently adds their second mobile phone number to the
corporate directory.

Handle this situation more gracefully by making an arbitrary choice to
use the first attribute value.  While nothing requires LDAP servers to
maintain the order of attribute values, it appears that at least
FreeIPA does so, and the order reflects the order as shown in the web
user interface.  Choosing the first listed value should therefore be
minimally surprising in the common case of a FreeIPA source database.

We extend this handling further by allow the destination attribute to
retain its current value if that value is one of the current source
attribute values.  This ensures that the arbitrary choice to use the
first attribute value has to be made only once.  On subsequent
synchronizations with an unchanged source attribute value list, the
destination attribute value will remain unchanged even if the source
attribute values have been reordered.  We therefore avoid requiring
that the source database maintains the order of attribute values,
since our destination attribute value will remain stable even when the
source order is randomised.

As a pleasant side effect, this allows end users to choose to edit the
destination attribute value to reflect a personal preference between
the possible source attribute values.  For example: a user may record
two mobile phone numbers in the corporate FreeIPA database.  By
default, the first listed number will be synchronized into a
single-phone-number database such as RT.  The user may choose to edit
RT to use their second listed number: the synchronization process will
not overwrite this choice (provided that the second number remains
listed in the FreeIPA source database).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>