ldap: Provide ability to trace and log database changes
The standard ldapsearch tool does not capture enough information to
replay an LDAP search in detail.  In particular, it does not capture
the value of intermediate results, which carry the syncInfoMessage
values.

Add a tool "idiotrace" which can be used to generate an LDIF file
incorporating all of the required information.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>