base: Provide a default implementation of find_syncid()
There is no need to force all database entry subclasses to provide an
implementation of find_syncid(), since it can trivially be implemented
as a special case of find_syncids().

Provide a default implementation so that subclasses may choose to
implement find_syncid() only if a dedicated single-syncid lookup
provides a useful efficiency gain.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>