base: Define a per-class Self type variable where applicable
The recommended pattern for annotating class methods that return
instances (e.g. alternative constructors) involves first defining a
type variable bound to the class and then explicitly annotating the
"cls" parameter of the class method.  This can become excessively
verbose given that type variables may not be reassigned and so require
unique names in the global namespace.

As a cleaner alternative, define a type variable "Self" within the
class namespace, thereby allowing alternative constructors to be
annotated using syntax such as:

  def create(cls: Type[Self]) -> Self: ...

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>